Moved Middlesection from app.js to own file
